Refactor LeadsTable to functional component using typescript and hooks
